cards edit

+++ src/main/java/card/uz/cards/dto/CardCreateDTO.java
@@ -14,6 +14,6 @@ import java.math.BigDecimal;
 public class CardCreateDTO {
     private Long user_id;
     private CardStatus status;
-    private BigDecimal balance;
+    private Long balance;
     private Currency currency;
 }
